Reorganize lifting/__init__.py.
